Move taxonomy route & template to theme
